blockthespot: Only replace chrome_elf.dll if required

+++ scripts/blockthespot.ps1
@@ -17,11 +17,14 @@ if ((Split-Path $spotify_dir_parent -leaf) -ne "spotify-latest") {
     }
 }
 
-$spotify_running = Get-Process -ErrorAction Ignore -Name Spotify
-Stop-Process -ErrorAction Ignore -Name Spotify | Out-Null
+if ((Get-FileHash "$spotify_dir\chrome_elf.dll").Hash -ne (Get-FileHash "$PSScriptRoot\chrome_elf.dll").Hash) {
+    $spotify_running = Get-Process -ErrorAction Ignore -Name Spotify
+    Stop-Process -ErrorAction Ignore -Name Spotify | Out-Null
 
-Move-Item -Force "$spotify_dir\chrome_elf.dll" -Destination "$spotify_dir\chrome_elf.dll.original"
-Copy-Item "$PSScriptRoot\chrome_elf.dll" -Destination "$spotify_dir"
-if (-not (Get-Content -ErrorAction Ignore "$spotify_dir\config.ini")) { Copy-Item "$PSScriptRoot\config.ini" -Destination "$spotify_dir" }
+    Move-Item -Force "$spotify_dir\chrome_elf.dll" -Destination "$spotify_dir\chrome_elf.dll.original"
+    Copy-Item "$PSScriptRoot\chrome_elf.dll" -Destination "$spotify_dir"
 
-if ($spotify_running) { Start-Process "$spotify_path" }
+    if (-not (Get-Content -ErrorAction Ignore "$spotify_dir\config.ini")) { Copy-Item "$PSScriptRoot\config.ini" -Destination "$spotify_dir" }
+
+    if ($spotify_running) { Start-Process "$spotify_path" }
+}
